+++ docker-images/kafka-based/kafka/scripts/kafka_tls_prepare_certificates.sh
@@ -19,11 +19,6 @@ for CERT_DIR in /opt/kafka/certificates/*; do
   fi
 done
 
-if [ -n "$STRIMZI_OPA_AUTHZ_TRUSTED_CERTS" ]; then
-  echo "Preparing Open Policy Agent authorization truststore"
-  prepare_truststore "/tmp/kafka/authz-opa.truststore.p12" "$CERTS_STORE_PASSWORD" "/opt/kafka/certificates/authz-opa-certs" "$STRIMZI_OPA_AUTHZ_TRUSTED_CERTS"
-fi
-
 if [ -n "$STRIMZI_KEYCLOAK_AUTHZ_TRUSTED_CERTS" ]; then
   echo "Preparing Keycloak authorization truststore"
   prepare_truststore "/tmp/kafka/authz-keycloak.truststore.p12" "$CERTS_STORE_PASSWORD" "/opt/kafka/certificates/authz-keycloak-certs" "$STRIMZI_KEYCLOAK_AUTHZ_TRUSTED_CERTS"