+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ### Major changes, deprecations, and removals
 
-* n/a
+* Open Policy Agent (OPA) authorizer plugin is not bundled as part of the Strimzi Container images anymore.
+  If you want to continue use the OPA Authorizer, you have to add it as a custom plugin by building a custom Kafka container image or using additional volumes.
+  Once added, you can continue to use OPA using the `type: custom` authorization.
 
 ## 0.51.0